Add ability to write to PCAP
This commit adds the ability for users to store unexpected packets in a PCAP file. This leverages the "EXPORT" environmental variable, which must be set in the Docker container on execution. The "EXPORT" environmental variable should be set to a filename, most likely on the bootflash of the device. Furthermore, if this environmental variable is used, the /bootflash directory should be directly mounted as a volume into the Docker container.